Update: Refine Qualcomm HWRNG test script and docs
Revised script and documentation following review suggestions.

+++ Runner/suites/Kernel/FunctionalArea/baseport/qcom_hwrng/run.sh
@@ -1,24 +1,31 @@
+#!/bin/sh
 # Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
 # SPDX-License-Identifier: BSD-3-Clause-Clear
 
-#!/bin/sh
 # Import test suite definitions
-. $(pwd)/init_env
+. "${PWD}"/init_env
 TESTNAME="qcom_hwrng"
 
 #import test functions library
-. $TOOLS/functestlib.sh
+. "${TOOLS}"/functestlib.sh
 test_path=$(find_test_case_by_name "$TESTNAME")
 log_info "-----------------------------------------------------------------------------------------"
 log_info "-------------------Starting $TESTNAME Testcase----------------------------"
 
 # Set the hardware RNG source to Qualcomm's RNG
-echo qcom_hwrng > /sys/class/misc/hw_random/rng_current
+if [ -e /sys/class/misc/hw_random/rng_current ]; then
+    echo qcom_hwrng > /sys/class/misc/hw_random/rng_current
+else
+    echo "Path /sys/class/misc/hw_random/rng_current does not exist."
+    log_fail "$TESTNAME : Test Failed"
+    exit 1
+fi
 
 # Verify that qcom_hwrng was successfully set
 current_rng=$(cat /sys/class/misc/hw_random/rng_current)
 if [ "$current_rng" != "qcom_hwrng" ]; then
     log_info "Error: Failed to set qcom_hwrng as the current RNG source."
+    log_fail "$TESTNAME : Test Failed"
     exit 1
 else
     log_info "qcom_hwrng successfully set as the current RNG source."
@@ -37,7 +44,7 @@ if [ "$value" -lt 10 ]; then
     log_pass "$TESTNAME : Test Passed"
     echo "$TESTNAME PASS" > $test_path/$TESTNAME.res
 else
-	log_fail "$TESTNAME : Test Failed"
-	echo "$TESTNAME FAIL" > $test_path/$TESTNAME.res
+    log_fail "$TESTNAME : Test Failed"
+    echo "$TESTNAME FAIL" > $test_path/$TESTNAME.res
 fi
 log_info "-------------------Completed $TESTNAME Testcase----------------------------"
